tidied up the extension methods, and added more to the InitialHandshakeHandler class

+++ simple-server/src/me/mdbell/terranet/server/simple/SimpleAttributes.java
@@ -12,5 +12,5 @@ public class SimpleAttributes implements ConnectionAttributes {
     String uuid;
     int currentHp, maxHp;
     int currentMana, maxMana;
-    final byte[] buffs = new byte[Opcodes.MAX_BUFFS];
+    final int[] buffs = new int[Opcodes.MAX_BUFFS];
 }
